Rewrite gender formatting
