+++ apps/judicial-system/backend/src/app/formatters/index.ts
@@ -8,5 +8,4 @@ export {
   formatCourtDateNotificationCondition,
   formatPrisonRulingEmailNotification,
   stripHtmlTags,
-  formatGender,
 } from './formatters'
